refactor: update platform API import paths to use local library
