+++ apps/electron/package.json
@@ -24,7 +24,6 @@
   },
   "dependencies": {
     "@ai-sdk/react": "^1.2.12",
-    "@mcp-router/platform-api": "workspace:*",
     "@mcp-router/shared": "workspace:*",
     "@mcp-router/tailwind-config": "workspace:*",
     "@mcp-router/ui": "workspace:*",
