+++ apps/electron/src/frontend/components/agent/AgentAuthGuard.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useAuthStore } from "@/frontend/stores";
-import { usePlatformAPI } from "@mcp-router/platform-api";
+import { usePlatformAPI } from "@/lib/platform-api";
 import LoginScreen from "../setup/LoginScreen";
 
 interface AgentAuthGuardProps {
