+++ apps/electron/src/frontend/components/agent/BackgroundComponent.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState, useRef } from "react";
 import { useChat, Message } from "@ai-sdk/react";
 import { AgentConfig } from "@mcp-router/shared";
 import { getServerAgentId } from "@/lib/utils/agent-utils";
-import { usePlatformAPI } from "@mcp-router/platform-api";
+import { usePlatformAPI } from "@/lib/platform-api";
 
 interface BackgroundComponentProps {
   chatHistorySessionId?: string; // チャット履歴のsessionId
