+++ apps/electron/src/frontend/components/agent/PackageManagerOverlay.tsx
@@ -16,7 +16,7 @@ import {
 } from "lucide-react";
 import { toast } from "sonner";
 import { useTranslation } from "react-i18next";
-import { usePlatformAPI } from "@mcp-router/platform-api";
+import { usePlatformAPI } from "@/lib/platform-api";
 
 interface PackageManagerOverlayProps {
   isOpen: boolean;
