+++ apps/electron/src/frontend/components/agent/create/AgentBuild.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useCallback } from "react";
 import { useNavigate } from "react-router-dom";
 import { AgentConfig } from "@mcp-router/shared";
-import { usePlatformAPI } from "@mcp-router/platform-api";
+import { usePlatformAPI } from "@/lib/platform-api";
 import { Button } from "@mcp-router/ui";
 import {
   Card,
