+++ apps/electron/src/frontend/components/mcp/rules/Rules.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { usePlatformAPI } from "@mcp-router/platform-api";
+import { usePlatformAPI } from "@/lib/platform-api";
 import { useWorkspaceStore } from "../../../stores";
 import {
   MCPDisplayRules,
