+++ apps/electron/src/frontend/stores/auth-store.ts
@@ -1,6 +1,6 @@
 import { create, StoreApi, UseBoundStore } from "zustand";
 import { AppSettings } from "@mcp-router/shared";
-import { PlatformAPI } from "@mcp-router/platform-api";
+import { PlatformAPI } from "@/lib/platform-api";
 
 interface UserInfo {
   userId: string;
