+++ apps/electron/src/lib/platform-api/types/domains/agent-api.ts
