+++ apps/electron/src/lib/platform-api/types/domains/app-api.ts
