+++ apps/electron/src/lib/platform-api/types/domains/auth-api.ts
