+++ apps/electron/src/lib/platform-api/types/domains/log-api.ts
