+++ apps/electron/src/lib/platform-api/types/domains/server-api.ts
