+++ apps/electron/src/lib/platform-api/types/domains/settings-api.ts
