+++ apps/electron/src/lib/platform-api/types/platform-api.ts
