+++ apps/electron/src/lib/platform-api/types/window.d.ts
