--- packages/platform-api/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@mcp-router/platform-api",
-  "version": "0.5.1",
-  "description": "Platform API abstraction for MCP Router",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "clean": "rimraf dist",
-    "lint": "eslint . --ext .ts,.tsx",
-    "type-check": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@mcp-router/shared": "workspace:*",
-    "react": "^19.0.0"
-  },
-  "devDependencies": {
-    "@types/react": "^19.0.7",
-    "eslint": "^9.0.0",
-    "rimraf": "^6.0.1",
-    "typescript": "^5.8.3"
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    }
-  }
-}
