--- packages/platform-api/src/index.ts
@@ -1,16 +1,0 @@
-// Platform API interfaces
-export { PlatformAPI } from "./platform-api-interface";
-export * from "./types/domains/auth-api";
-export * from "./types/domains/server-api";
-export * from "./types/domains/agent-api";
-export * from "./types/domains/app-api";
-export * from "./types/domains/package-api";
-export * from "./types/domains/settings-api";
-export * from "./types/domains/log-api";
-export * from "./types/domains/workspace-api";
-
-// Platform API factory and utilities
-export { isElectron, isWeb, createPlatformAPI } from "./platform-api-factory";
-
-// Platform API React context and provider
-export { PlatformAPIProvider, usePlatformAPI } from "./platform-api-context";
